ark-core/agents/composio_bridge/requirements.txt
nats-py==2.9.0
requests==2.32.3
